Rating calculation (#5)

+++ domain/food.go
@@ -1,9 +1,9 @@
 package domain
 
 type Food struct {
-	Id               int    `json:"id"`
-	Name             string `json:"name"`
-	PreparationTime  int    `json:"preparation-time"`
-	Complexity       int    `json:"complexity"`
-	CookingApparatus string `json:"cooking-apparatus"`
+	Id               int     `json:"id"`
+	Name             string  `json:"name"`
+	PreparationTime  float64 `json:"preparation-time"`
+	Complexity       int     `json:"complexity"`
+	CookingApparatus string  `json:"cooking-apparatus"`
 }
